Carve off AdvancedSuiteTest
