+++ corehq/apps/app_manager/tests/data/suite/app_graphing.json
@@ -33,7 +33,7 @@
            "logenabled": "Enabled"
        }
    },
-   "domain": "example",
+   "domain": "domain",
    "build_langs": [
        "en",
        "fra"
