+++ corehq/apps/app_manager/tests/data/suite/careplan.json
@@ -2,7 +2,7 @@
    "_id": "747fda5aaa48c3e9bceda6295db74475",
    "_rev": "21-0aa5dbf2678e8f76a7e7db83346136da",
    "short_odk_url": null,
-   "domain": "skelly",
+   "domain": "domain",
    "build_langs": [
        "en"
    ],
