+++ corehq/apps/app_manager/tests/data/suite/suite-advanced.json
@@ -2,7 +2,7 @@
   "_id": "5599cb34f0262fbf4796b599508f84ca",
   "_rev": "37-0f7a55e21778fe782a7ee041d5f85eb7",
   "short_odk_url": null,
-  "domain": "skelly",
+  "domain": "domain",
   "build_langs": [
     "en"
   ],
