+++ corehq/apps/app_manager/tests/test_suite.py
@@ -3,18 +3,11 @@ import re
 from django.test import SimpleTestCase
 from corehq.apps.app_manager.const import APP_V2
 from corehq.apps.app_manager.models import (
-    AUTO_SELECT_CASE,
-    AUTO_SELECT_FIXTURE,
-    AUTO_SELECT_RAW,
-    AUTO_SELECT_USER,
-    AUTO_SELECT_USERCASE,
     AdvancedModule,
     Application,
-    AutoSelectCase,
     DetailColumn,
     FormActionCondition,
     FormSchedule,
-    LoadUpdateAction,
     MappingItem,
     Module,
     OpenCaseAction,
@@ -27,7 +20,7 @@ from corehq.apps.app_manager.models import (
     UpdateCaseAction,
 )
 from corehq.apps.app_manager.tests.app_factory import AppFactory
-from corehq.apps.app_manager.tests.util import TestXmlMixin, commtrack_enabled
+from corehq.apps.app_manager.tests.util import SuiteMixin, TestXmlMixin, commtrack_enabled
 from corehq.apps.app_manager.xpath import (
     session_var,
 )
@@ -35,52 +28,15 @@ from corehq.toggles import NAMESPACE_DOMAIN
 from corehq.feature_previews import MODULE_FILTER
 from toggle.shortcuts import update_toggle_cache, clear_toggle_cache
 
-from lxml import etree
-import commcare_translations
 
-
-class SuiteTest(SimpleTestCase, TestXmlMixin):
+class SuiteTest(SimpleTestCase, TestXmlMixin, SuiteMixin):
     file_path = ('data', 'suite')
 
     def setUp(self):
-        update_toggle_cache(MODULE_FILTER.slug, 'skelly', True, NAMESPACE_DOMAIN)
         update_toggle_cache(MODULE_FILTER.slug, 'domain', True, NAMESPACE_DOMAIN)
-        update_toggle_cache(MODULE_FILTER.slug, 'example', True, NAMESPACE_DOMAIN)
 
     def tearDown(self):
-        clear_toggle_cache(MODULE_FILTER.slug, 'skelly', NAMESPACE_DOMAIN)
         clear_toggle_cache(MODULE_FILTER.slug, 'domain', NAMESPACE_DOMAIN)
-        clear_toggle_cache(MODULE_FILTER.slug, 'example', NAMESPACE_DOMAIN)
-
-    def assertHasAllStrings(self, app, strings):
-        et = etree.XML(app)
-        locale_elems = et.findall(".//locale/[@id]")
-        locale_strings = [elem.attrib['id'] for elem in locale_elems]
-
-        app_strings = commcare_translations.loads(strings)
-
-        for string in locale_strings:
-            if string not in app_strings:
-                raise AssertionError("App strings did not contain %s" % string)
-            if not app_strings.get(string, '').strip():
-                raise AssertionError("App strings has blank entry for %s" % string)
-
-    def _test_generic_suite(self, app_tag, suite_tag=None):
-        suite_tag = suite_tag or app_tag
-        app = Application.wrap(self.get_json(app_tag))
-        self.assertXmlEqual(self.get_xml(suite_tag), app.create_suite())
-
-    def _test_generic_suite_partial(self, app_tag, xpath, suite_tag=None):
-        suite_tag = suite_tag or app_tag
-        app = Application.wrap(self.get_json(app_tag))
-        self.assertXmlPartialEqual(self.get_xml(suite_tag), app.create_suite(), xpath)
-
-    def _test_app_strings(self, app_tag):
-        app = Application.wrap(self.get_json(app_tag))
-        app_xml = app.create_suite()
-        app_strings = app.create_app_strings('default')
-
-        self.assertHasAllStrings(app_xml, app_strings)
 
     def test_normal_suite(self):
         self._test_generic_suite('app', 'normal-suite')
@@ -125,36 +81,6 @@ class SuiteTest(SimpleTestCase, TestXmlMixin):
         app.get_module(1).display_separately = True
         self.assertXmlEqual(self.get_xml('careplan-own-module'), app.create_suite())
 
-    def test_advanced_suite(self):
-        self._test_generic_suite('suite-advanced')
-
-    def test_advanced_suite_details(self):
-        app = Application.wrap(self.get_json('suite-advanced'))
-        clinic_module_id = app.get_module(0).unique_id
-        other_module_id = app.get_module(1).unique_id
-        app.get_module(1).get_form(0).actions.load_update_cases[0].details_module = clinic_module_id
-        app.get_module(1).get_form(1).actions.load_update_cases[0].details_module = other_module_id
-        self.assertXmlEqual(self.get_xml('suite-advanced-details'), app.create_suite())
-
-    def test_advanced_suite_parent_child_custom_ref(self):
-        app = Application.wrap(self.get_json('suite-advanced'))
-        form = app.get_module(1).get_form(2)
-        form.actions.load_update_cases[1].case_index.reference_id = 'custom-parent-ref'
-        self.assertXmlPartialEqual(self.get_xml('custom-parent-ref'), app.create_suite(), "./entry[4]")
-
-    def test_advanced_suite_case_list_filter(self):
-        app = Application.wrap(self.get_json('suite-advanced'))
-        clinic_module = app.get_module(0)
-        clinic_module.case_details.short.filter = "(filter = 'danny')"
-        clinic_module_id = clinic_module.unique_id
-        app.get_module(1).get_form(0).actions.load_update_cases[0].details_module = clinic_module_id
-        self.assertXmlEqual(self.get_xml('suite-advanced-filter'), app.create_suite())
-
-    @commtrack_enabled(True)
-    def test_advanced_suite_commtrack(self):
-        app = Application.wrap(self.get_json('suite-advanced'))
-        self.assertXmlEqual(self.get_xml('suite-advanced-commtrack'), app.create_suite())
-
     @commtrack_enabled(True)
     def test_product_list_custom_data(self):
         # product data shouldn't be interpreted as a case index relationship
@@ -181,89 +107,6 @@ class SuiteTest(SimpleTestCase, TestXmlMixin):
             app_xml,
             './entry[1]'
         )
-
-    def test_advanced_suite_auto_select_user(self):
-        app = Application.wrap(self.get_json('suite-advanced'))
-        app.get_module(1).get_form(0).actions.load_update_cases[0].auto_select = AutoSelectCase(
-            mode=AUTO_SELECT_USER,
-            value_key='case_id'
-        )
-        self.assertXmlPartialEqual(self.get_xml('suite-advanced-autoselect-user'), app.create_suite(),
-                                   './entry[2]')
-
-    def test_advanced_suite_auto_select_fixture(self):
-        app = Application.wrap(self.get_json('suite-advanced'))
-        app.get_module(1).get_form(0).actions.load_update_cases[0].auto_select = AutoSelectCase(
-            mode=AUTO_SELECT_FIXTURE,
-            value_source='table_tag',
-            value_key='field_name'
-        )
-        self.assertXmlPartialEqual(self.get_xml('suite-advanced-autoselect-fixture'), app.create_suite(),
-                                   './entry[2]')
-
-    def test_advanced_suite_auto_select_raw(self):
-        app = Application.wrap(self.get_json('suite-advanced'))
-        app.get_module(1).get_form(0).actions.load_update_cases[0].auto_select = AutoSelectCase(
-            mode=AUTO_SELECT_RAW,
-            value_key=("some xpath expression "
-                       "containing instance('casedb') "
-                       "and instance('commcaresession')")
-        )
-        self.assertXmlPartialEqual(self.get_xml('suite-advanced-autoselect-raw'), app.create_suite(),
-                                   './entry[2]')
-
-    def test_advanced_suite_auto_select_case(self):
-        app = Application.wrap(self.get_json('suite-advanced'))
-        load_update_cases = app.get_module(1).get_form(0).actions.load_update_cases
-        load_update_cases.append(LoadUpdateAction(
-            case_tag='auto_selected',
-            auto_select=AutoSelectCase(
-                mode=AUTO_SELECT_CASE,
-                value_source=load_update_cases[0].case_tag,
-                value_key='case_id_index'
-            )
-        ))
-        self.assertXmlPartialEqual(self.get_xml('suite-advanced-autoselect-case'), app.create_suite(),
-                                   './entry[2]')
-
-    def test_advanced_suite_auto_select_usercase(self):
-        app = Application.wrap(self.get_json('suite-advanced'))
-        app.get_module(1).get_form(0).actions.load_update_cases[0].auto_select = AutoSelectCase(
-            mode=AUTO_SELECT_USERCASE
-        )
-        self.assertXmlPartialEqual(self.get_xml('suite-advanced-autoselect-usercase'), app.create_suite(),
-                                   './entry[2]')
-
-    def test_advanced_suite_auto_select_with_filter(self):
-        """
-        Form filtering should be done using the last 'non-autoload' case being loaded.
-        """
-        app = Application.wrap(self.get_json('suite-advanced'))
-        app.get_module(1).get_form(0).actions.load_update_cases.append(LoadUpdateAction(
-            case_tag='autoload',
-            auto_select=AutoSelectCase(
-                mode=AUTO_SELECT_USER,
-                value_key='case_id'
-            )
-        ))
-        form = app.get_module(1).get_form(0)
-        form.form_filter = "./edd = '123'"
-        suite = app.create_suite()
-        self.assertXmlPartialEqual(self.get_xml('suite-advanced-autoselect-with-filter'), suite, './entry[2]')
-        menu = """
-        <partial>
-          <menu id="m1">
-            <text>
-              <locale id="modules.m1"/>
-            </text>
-            <command id="m1-f0" relevant="instance('casedb')/casedb/case[@case_id=instance('commcaresession')/session/data/case_id_case_clinic]/edd = '123'"/>
-            <command id="m1-f1"/>
-            <command id="m1-f2"/>
-            <command id="m1-case-list"/>
-          </menu>
-        </partial>
-        """
-        self.assertXmlPartialEqual(menu, suite, "./menu[@id='m1']")
 
     def test_case_assertions(self):
         self._test_generic_suite('app_case_sharing', 'suite-case-sharing')
@@ -349,26 +192,6 @@ class SuiteTest(SimpleTestCase, TestXmlMixin):
             app.create_suite(),
             "./entry[1]"
         )
-
-    def test_tiered_select_with_advanced_module_as_parent(self):
-        app = Application.new_app('domain', "Untitled Application", application_version=APP_V2)
-
-        parent_module = app.add_module(AdvancedModule.new_module('parent', None))
-        parent_module.case_type = 'parent'
-        parent_module.unique_id = 'id_parent_module'
-
-        child_module = app.add_module(Module.new_module("Untitled Module", None))
-        child_module.case_type = 'child'
-        child_module.parent_select.active = True
-
-        # make child module point to advanced module as parent
-        child_module.parent_select.module_id = parent_module.unique_id
-
-        child_form = app.new_form(1, "Untitled Form", None)
-        child_form.xmlns = 'http://id_m1-f0'
-        child_form.requires = 'case'
-
-        self.assertXmlPartialEqual(self.get_xml('advanced_module_parent'), app.create_suite(), "./entry[1]")
 
     def test_usercase_id_added_update(self):
         app = Application.new_app('domain', "Untitled Application", application_version=APP_V2)
